remove enum, fewer messages, simple survey
