+++ backend/app/main/answer_views.py
@@ -3,7 +3,8 @@ from twilio.twiml.messaging_response import MessagingResponse
 
 from . import main
 from .. import db
-from .models import Answer, Question, EnumResponse, TYPE_OBJECTS
+from .models import Answer, Question
+from .response_types import TYPE_OBJECTS
 
 
 @main.route('/answer/<question_id>', methods=['POST'])
@@ -14,13 +15,8 @@ def answer(question_id):
     # If not, prompt user
     if not is_allowed_answer(request.values['Body'], question):
         return redirect_invalid_twiml(question)
-	
-    if question.kind is Question.ENUM:
-        answer_body = EnumResponse(question).map_value(request.values['Body'])
-    else:
-        answer_body = request.values['Body']
 
-    db.save(Answer(content=answer_body,
+    db.save(Answer(content=request.values['Body'],
             question=question,
             session_id=session['id']))
 
@@ -33,10 +29,7 @@ def answer(question_id):
 
 # Verify that the answer matches the expected type
 def is_allowed_answer(body, question):
-    if question.kind is Question.ENUM:
-        return EnumResponse(question).is_of_type(body)
-    else:
-        return TYPE_OBJECTS[question.kind].is_of_type(body)
+    return TYPE_OBJECTS[question.kind].is_valid(body)
 
 
 # Redirect invalid answer
