+++ backend/app/main/models.py
@@ -1,8 +1,6 @@
 import json
 from app import db
-from abc import ABC, abstractmethod
 from werkzeug.security import generate_password_hash, check_password_hash
-import re
 
 # create model classes here
 
@@ -26,27 +24,21 @@ class Question(db.Model):
 
     TEXT = 'text'
     EMAIL = 'email'
-    ENUM = 'enum'
 
     id = db.Column(db.Integer, primary_key=True)
     content = db.Column(db.String, nullable=False)
-    kind = db.Column(db.Enum(TEXT, EMAIL, ENUM, name='question_kind'))
-    enum_values = db.Column(db.PickleType())
+    kind = db.Column(db.Enum(TEXT, EMAIL, name='question_kind'))
     survey_id = db.Column(db.Integer, db.ForeignKey('surveys.id'))
     answers = db.relationship('Answer', backref='question', lazy='dynamic')
 
-    def __init__(self, content, kind=TEXT, enum_values=[]):
+    def __init__(self, content, kind=TEXT):
         self.content = content
         self.kind = kind
-        self.enum_values = enum_values
 
     def next(self):
         return self.survey.questions\
 	            .filter(Question.id > self.id)\
                     .order_by('id').first()
-
-    def get_enum_values(self):
-        return self.enum_values
 
 
 class Answer(db.Model):
@@ -69,58 +61,3 @@ class Answer(db.Model):
         self.content = content
         self.question = question
         self.session_id = session_id
-
-
-class ResponseType(ABC):    
-
-    @abstractmethod
-    def is_of_type(self, arg):
-        pass
-    
-    @abstractmethod
-    def instruction_text(self):
-        pass
-
-
-class TextResponse(ResponseType):
-    def is_of_type(self, arg):
-        return True
-
-    def instruction_text(self):
-        return 'Please type your answer.'
-
-
-class EmailResponse(ResponseType):
-    def is_of_type(self, arg):
-        return re.match(r'[^@]+@[^@]+\.[^@]+', arg)
-
-    def instruction_text(self):
-        return 'Please type your answer.'
-
-
-class EnumResponse(ResponseType):
-    def __init__(self, question=None):
-        self.question = question
-
-    def is_of_type(self, arg):
-        try:
-            if self.map_value(arg) is not None:
-                return True
-            else:
-                return False
-        except (TypeError, IndexError):
-            return False
-
-    def instruction_text(self):
-        return 'Please enter the corresponding number.'
-
-    def map_value(self, body):
-        return (self.question.get_enum_values())[int (body)]
-
-# Connect db types with their corresponding objects
-# for SMS instructions and validating responses
-TYPE_OBJECTS = {
-    Question.TEXT: TextResponse(),
-    Question.EMAIL: EmailResponse(),
-    Question.ENUM: EnumResponse()
-}
