+++ backend/app/main/question_views.py
@@ -2,7 +2,7 @@ from flask import session
 from twilio.twiml.messaging_response import MessagingResponse
 
 from . import main
-from .models import Question, TYPE_OBJECTS
+from .models import Question
 
 # Get the next question
 @main.route('/question/<question_id>', methods=['GET'])
@@ -12,10 +12,8 @@ def question(question_id):
     return sms_twiml(question)
 
 
-
 # Generate SMS messages for the question text and instructions for responding
 def sms_twiml(question):
     response = MessagingResponse()
     response.message(question.content)
-    response.message(TYPE_OBJECTS[question.kind].instruction_text())
     return str(response)
